fix: Moved test viewModel to it's project
