+++ src/CSharp/Blazor/EasyMicroservices.UI.Identity.Blazor.TestUI/ViewModels/ForgotPasswordChildViewModel.cs
@@ -3,12 +3,13 @@ using EasyMicroservices.ServiceContracts;
 using EasyMicroservices.UI.Cores;
 using EasyMicroservices.UI.Cores.Commands;
 using EasyMicroservices.UI.Identity.Models;
+using EasyMicroservices.UI.Identity.ViewModels.Authentications;
 using Identity.GeneratedServices;
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Text;
 
-namespace EasyMicroservices.UI.Identity.ViewModels.Authentications
+namespace EasyMicroservices.UI.Identity.Blazor.TestUI.ViewModels
 {
     public class ForgotPasswordChildViewModel : ForgotPasswordViewModel
     {
